chore: fix dependency inversion problems in controllers Here I was contemplating between the two choices: 1) Abstract InputHandler into Controller class and pass it there via super() in each subclass constructor 2) Abstract InputHandler into Controller and get rid of dependency inversion
I decided to go for the second option as it ends up in a less complicated and more loosely coupled code. Since InputHandler is a singleton with no variations planned for it in the future, the dependency inversion principle does not provide as many benefits as code simplification in this particular case.